Violet/requirements.txt
openai>=1.0.0
python-dotenv==1.0.0
pyttsx3==2.90
requests==2.31.0
SpeechRecognition==3.10.0
PyAudio==0.2.14
pywhatkit
numpy
fastapi
uvicorn
pydantic
google-genai
sounddevice